moved TextList
