+++ opencti-platform/opencti-front/src/components/common/text/TextList.tsx
