+++ opencti-platform/opencti-front/src/private/components/analyses/malware_analyses/MalwareAnalysisDetails.tsx
@@ -11,7 +11,7 @@ import FieldOrEmpty from '../../../../components/FieldOrEmpty';
 import Card from '../../../../components/common/card/Card';
 import Label from '../../../../components/common/label/Label';
 import StixDomainObjectNestedEntities from '@components/common/stix_domain_objects/StixDomainObjectNestedEntities';
-import TextList from '../../../../components/TextList';
+import TextList from '../../../../components/common/text/TextList';
 
 // Deprecated - https://mui.com/system/styles/basics/
 // Do not use it for new code.
