+++ opencti-platform/opencti-front/src/private/components/arsenal/malwares/MalwareDetails.jsx
@@ -13,7 +13,7 @@ import ItemBoolean from '../../../../components/ItemBoolean';
 import StixCoreObjectKillChainPhasesView from '../../common/stix_core_objects/StixCoreObjectKillChainPhasesView';
 import Label from '../../../../components/common/label/Label';
 import StixDomainObjectNestedEntities from '@components/common/stix_domain_objects/StixDomainObjectNestedEntities';
-import TextList from '../../../../components/TextList';
+import TextList from '../../../../components/common/text/TextList';
 
 const styles = (theme) => ({
   paper: {
