+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilityDetails.tsx
@@ -19,7 +19,7 @@ import Card from '../../../../components/common/card/Card';
 import Label from '../../../../components/common/label/Label';
 import Tag from '../../../../components/common/tag/Tag';
 import { Stack } from '@mui/material';
-import TextList from '../../../../components/TextList';
+import TextList from '../../../../components/common/text/TextList';
 
 const getCvssCriticity = (score: number | null | undefined): string | null => {
   if (typeof score !== 'number' || score < 0 || score > 10) return null;
